node_sensory/sonic_sensory.py
import os
import asyncio
import logging
import time
from cartesia import Cartesia
from cartesia.api import Stream
from typing import Dict, Any

# Assume mesh_firebase.py is in the parent 'utils' directory
import sys
sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
from utils.mesh_firebase import mesh_db

# --- Kaiju Noir Logging ---
logging.basicConfig(level=logging.INFO, format='%(asctime)s [SONIC SENSORY]: %(message)s')
logger = logging.getLogger("SonicSensory")

# --- Configuration ---
# Cartesia API Key must be set as an environment variable
CARTESIA_API_KEY = os.environ.get("CARTESIA_API_KEY")

class SonicSensoryHandler:
    """
    [ ⚡️ SONIC SENSORY HANDLER ]
    Achieves 100ms 'Speed Mouth' latency and full barge-in support by
    integrating the Cartesia (Sonic) Cloud API with the Ghidorah Mesh's
    Firestore C2 state machine.
    """

    def __init__(self):
        if not CARTESIA_API_KEY:
            logger.error("[FATAL] CARTESIA_API_KEY not found in environment. Sonic Sensory is offline.")
            raise ValueError("CARTESIA_API_KEY is not set.")
        
        self.cartesia_client = Cartesia(api_key=CARTESIA_API_KEY)
        self.is_streaming = False
        logger.info("Cartesia Sonic link established.")

    async def stream_tts(self, text_to_speak: str, voice_id: str = "79a3b672-5355-4471-b51e-35253401a935"):
        """
        Streams audio from Cartesia with <100ms latency while monitoring for interruptions.
        """
        self.is_streaming = True
        start_time = time.time()
        first_byte_latency = None

        # Report initial state to Firestore
        asyncio.create_task(mesh_db.update_node_status(
            node_id="node-sensory",
            role="sensory-output",
            status_data={"tts_status": "streaming", "current_text": text_to_speak}
        ))

        try:
            # This is a placeholder for your actual audio playback logic
            # In a real app, you'd use a library like `sounddevice` or `pyaudio`
            # and feed the chunks directly into the audio stream.
            async def mock_audio_playback(stream: Stream):
                nonlocal first_byte_latency
                async for chunk in stream:
                    if first_byte_latency is None:
                        first_byte_latency = time.time() - start_time
                        logger.info(f"First byte received in {first_byte_latency*1000:.2f}ms")
                    
                    # MOCK PLAYBACK: In a real app, you'd write chunk['audio'] to an audio device
                    await asyncio.sleep(0.05) # Simulate playback buffer time

                    # BARGE-IN DETECTION: Check for an external signal
                    # In a real app, this would come from a VAD (Voice Activity Detection)
                    # library monitoring the microphone.
                    if self._detect_barge_in():
                        logger.warning("[BARGE-IN DETECTED] User interruption. Halting TTS stream.")
                        # Signal Node C via Firestore
                        asyncio.create_task(mesh_db.update_node_status(
                            node_id="node-sensory",
                            role="sensory-output",
                            status_data={"interruption": True, "tts_status": "interrupted"}
                        ))
                        break # Stop iterating and end the stream

            # Get the audio stream from Cartesia
            audio_stream = self.cartesia_client.tts.stream(
                model_id="sonic-english",
                transcript=text_to_speak,
                voice_id=voice_id,
                output_format="fp32" # Raw PCM for lowest latency
            )
            
            await mock_audio_playback(audio_stream)

        except Exception as e:
            logger.error(f"Cartesia streaming failed: {e}")
            asyncio.create_task(mesh_db.update_node_status(
                node_id="node-sensory",
                role="sensory-output",
                status_data={"tts_status": "error", "error_message": str(e)}
            ))
        finally:
            self.is_streaming = False
            total_latency = time.time() - start_time
            logger.info(f"TTS stream finished. Total duration: {total_latency:.2f}s")
            
            # Report final state
            asyncio.create_task(mesh_db.update_node_status(
                node_id="node-sensory",
                role="sensory-output",
                status_data={
                    "tts_status": "idle", 
                    "voice_health": "nominal",
                    "last_stream_latency_ms": round(total_latency * 1000),
                    "first_byte_latency_ms": round(first_byte_latency * 1000) if first_byte_latency else -1
                }
            ))
            
    def _detect_barge_in(self) -> bool:
        """
        Mocks a barge-in detection mechanism. In a real application, this would
        interface with a Voice Activity Detection (VAD) library on the microphone input.
        """
        # For this simulation, we'll randomly trigger it.
        # import random
        # return random.random() < 0.01 
        return False # Keep it false for predictable testing

# --- Example Usage ---
async def main():
    try:
        handler = SonicSensoryHandler()
        text = "This is a test of the ultra-low latency Speed Mouth capability for the Ghidorah Mesh, ensuring full barge-in support and real-time telemetry."
        await handler.stream_tts(text)
    except ValueError as e:
        logger.error(f"Could not start handler: {e}")

if __name__ == "__main__":
    # Ensure you have set the CARTESIA_API_KEY environment variable
    asyncio.run(main())
