node_sensory/requirements.txt
faster-whisper
google-generativeai
firebase-admin
ffmpeg-python
